third_party/WebKit/LayoutTests/platform/win/virtual/outofblink-cors-http/http/tests/htmlimports/redirect-expected.txt

CONSOLE ERROR: Redirect from 'http://localhost:8000/htmlimports/resources/cors-basic.cgi' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Redirect from 'http://localhost:8000/htmlimports/resources/hello.html' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Accessing to a same origin import 
PASS Accessing to CORS-enabled origin import 
PASS Blocking access to a cross origin import 
PASS Accessing to a same origin import including same domain redirect 
FAIL Accessing to a CORS origin import including same domain redirect assert_true: expected true got false
PASS Blocking access to a cross origin iamport including same domain redirect 
Harness: the test ran to completion.
